backend/requirements.txt
alembic==1.16.5
aniso8601==10.0.1
blinker==1.9.0
boto3==1.40.25
botocore==1.40.25
click==8.2.1
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
s3transfer==0.13.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
pydantic==2.8.2
celery==5.4.0
redis==5.0.7